Revert "wip: integration test"
This reverts commit 9b8858036f49a846a6880dc67c382bc8511084ac.
