+++ crates/forge/Cargo.toml
@@ -103,7 +103,6 @@ tikv-jemallocator = { workspace = true, optional = true }
 [dev-dependencies]
 anvil.workspace = true
 foundry-test-utils.workspace = true
-alloy-contract.workspace = true
 
 mockall = "0.13"
 globset = "0.4"
