+++ crates/forge/src/cmd/create.rs
@@ -421,9 +421,15 @@ impl CreateArgs {
             // self.naming.run("deployandname").await?;
             let config = self.load_config()?;
             let signer = self.eth.wallet.signer().await?;
+            let provider = utils::get_provider(&config)?;
+            let provider = ProviderBuilder::<_, _, AnyNetwork>::default()
+                .with_recommended_fillers()
+                .wallet(EthereumWallet::new(signer))
+                .on_provider(provider);
+            let sender_addr = provider.default_signer_address();
             enscribe::set_primary_name(
-                &config,
-                EthereumWallet::new(signer),
+                provider,
+                sender_addr,
                 deployed_contract,
                 self.ens_name,
                 self.reverse_claimer,
