--- crates/forge/tests/cli/abi.rs
@@ -1,7 +1,0 @@
-use alloy_sol_types::sol;
-
-sol!(
-    #[sol(rpc)]
-    ENSRegistry,
-    "test-data/ENSRegistry.json"
-);
