+++ crates/forge/tests/cli/main.rs
@@ -29,6 +29,5 @@ mod verify;
 mod verify_bytecode;
 mod version;
 
-mod abi;
 mod ext_integration;
 mod test_optimizer;
